two commit

+++ src/BmiService.java
@@ -1,7 +1,7 @@
 public class BmiService {
-    public double calculate (int weightKg, double heightM) {
+    public int calculate (int weightKg, double heightM) {
         double result;
         result = (weightKg / (heightM * heightM));
-        return result;
+        return (int) result;
     }
 }
